+++ webhookService/domain/identificationResult.go
@@ -14,7 +14,7 @@ type IdentificationResult struct {
 	Organization string      `json:"organization"`
 	Department   string      `json:"department"`
 	TicketID     string      `json:"ticketID"`
-	BoundingBox  [4]int      `json:"boundingBox"`
+	BoundingBox  [4]float64  `json:"boundingBox"`
 	Timestamp    uint64      `json:"timestamp"`
 	Metadata     interface{} `json:"metadata"`
 }
